+++ Core/src/main/java/com/gitlab/pauloo27/core/sql/EzSQL.java
@@ -23,7 +23,8 @@ public abstract class EzSQL<DatabaseType extends Database, TableType extends Tab
      */
     private static DataType DEFAULT_DATATYPE = new DataType(
             "VARCHAR",
-            DefaultAttributes.getDefaultAttributes(),
+            true,
+            "64",
             null,
             "DEFAULT_DATATYPE"
     );
@@ -32,7 +33,8 @@ public abstract class EzSQL<DatabaseType extends Database, TableType extends Tab
      */
     private static DataType ENUM_NAME = new DataType(
             "VARCHAR",
-            DefaultAttributes.getDefaultAttributes(),
+            true,
+            "64",
             null,
             "ENUM_NAME"
     );
@@ -802,7 +804,7 @@ public abstract class EzSQL<DatabaseType extends Database, TableType extends Tab
                 tableBuilder.withColumn(
                         new ColumnBuilder(columnName, dataType, length, attributes.toArray(new Attribute[]{}))
                                 .withDefaultValue(defaultValue));
-            } catch (SQLException | InstantiationException | IllegalAccessException e) {
+            } catch (InstantiationException | IllegalAccessException e) {
                 e.printStackTrace();
             }
         });